+++ models/invitees.py
@@ -6,4 +6,4 @@ class Invitees(db.Model, UserMixin):
     id = db.Column(db.Integer, primary_key=True)
     first_name = db.Column(db.String(120))
     last_name = db.Column(db.String(120))
-    invited_by = db.Column(db.Integer, db.ForeignKey("user.id"))
+    invited_by = db.Column(db.BigInteger, db.ForeignKey("user.id"))
